scraper/requirements.txt
httpx>=0.27
selectolax>=0.3
asyncpg>=0.30
python-dotenv>=1.0
